new/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dateutil